+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Added
 
 - `Innmind\Immutable\Set::match()`
+- `Innmind\Immutable\Predicate\OrPredicate`
+- `Innmind\Immutable\Predicate\AndPredicate`
+- `Innmind\Immutable\Predicate\Instance::or()`
+- `Innmind\Immutable\Predicate\Instance::and()`
 
 ## 5.1.0 - 2023-10-11